getProjectConfig extensions hijacks `toolbox.config`
Currently the `getProjectConfig` extension, which fetches
devctl monorepo config simple overwrites `toolbox.config`
which causes problems with plugins wanting to expose
configuration on that same object either by extending it on its own,
or by using gluegun's built-in cosmiconfig.

Fix is to destructure the previous contents of `toolbox.config`
before it's mutated.